final project

+++ lib/data/notifier.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 ValueNotifier<String> genderValueNotifier = ValueNotifier("Null");
 
-ValueNotifier<List<ValueNotifier<int>>> weightAndAgeValueNotifier =
+ValueNotifier<List<ValueNotifier<double>>> weightAndAgeValueNotifier =
     ValueNotifier([ValueNotifier(60), ValueNotifier(20)]); // Weight - Age
 
-ValueNotifier<int> heightValueNotifier = ValueNotifier(100);
+ValueNotifier<double> heightValueNotifier = ValueNotifier(100);
